Moved inference prov functions to different files
