+++ modules/lib_vnr/api/inference/vnr_inference_api.h
@@ -10,8 +10,6 @@ extern "C" {
     #include "bfp_math.h"
     void vnr_inference_init(vnr_ie_state_t *ie_ptr);
     void vnr_inference(vnr_ie_state_t *ie_state, float_s32_t *vnr_output, bfp_s32_t *features);
-    void vnr_priv_output_dequantise(float_s32_t *dequant_output, const int8_t* quant_output, const vnr_ie_config_t *ie_config);
-    void vnr_priv_feature_quantise(int8_t *quantised_patch, bfp_s32_t *normalised_patch, const vnr_ie_config_t *ie_config);
 #ifdef __cplusplus
 }
 #endif
